+++ tests/base.py
@@ -100,3 +100,13 @@ class PyexcelMultipleSheetBase(TestCase):
         data = list(b["Sheet3"].rows())
         expected = [[1, 4, 7], [2, 5, 8], [3, 6, 9]]
         assert data == expected
+
+
+def _produce_ordered_dict():
+    data_dict = OrderedDict()
+    data_dict.update({"Sheet1": [[1, 1, 1, 1], [2, 2, 2, 2], [3, 3, 3, 3]]})
+    data_dict.update({"Sheet2": [[4, 4, 4, 4], [5, 5, 5, 5], [6, 6, 6, 6]]})
+    data_dict.update(
+        {"Sheet3": [["X", "Y", "Z"], [1, 4, 7], [2, 5, 8], [3, 6, 9]]}
+    )
+    return data_dict
